+++ cmd/tmux-project/constants.go
@@ -2,7 +2,6 @@ package main
 
 const (
 	ApplicationName string = "tmux-project"
-	Editor          string = "vim"
 	ConfigDirName   string = "tmux-project"
 	// ConstDirMode permissions on newly created configuration directories
 	ConstDirMode int = 0755
