+++ cmd/tmux-project/root.go
@@ -9,6 +9,7 @@ import (
 	"path"
 
 	"github.com/jvzantvoort/tmux-project/messages"
+	"github.com/mitchellh/go-homedir"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -69,7 +70,7 @@ func GetConfigParamName(instr string) string {
 
 func PrincipalConfigDir() string {
 	// Find home directory.
-	home, err := os.UserHomeDir()
+	home, err := homedir.Dir()
 	cobra.CheckErr(err)
 
 	return path.Join(home, ".config", ApplicationName)
