+++ config/main.go
@@ -1,11 +1,11 @@
 // Package config provides configuration data globally used
-//
 package config
 
 import (
-	"os/user"
 	"path"
-	"path/filepath"
+
+	"github.com/mitchellh/go-homedir"
+	"github.com/jvzantvoort/tmux-project/utils"
 )
 
 type MainConfig struct {
@@ -14,30 +14,13 @@ type MainConfig struct {
 	ProjTypeConfigDir string
 }
 
-// ExpandHome expand the tilde in a given path.
-func (m MainConfig) ExpandHome(pathstr string) (string, error) {
-	if len(pathstr) == 0 {
-		return pathstr, nil
-	}
-
-	if pathstr[0] != '~' {
-		return pathstr, nil
-	}
-
-	return filepath.Join(m.HomeDir, pathstr[1:]), nil
-
-}
-
 func NewMainConfig() *MainConfig {
 
 	v := &MainConfig{}
 
-	usr, err := user.Current()
-	if err != nil {
-		panic(err)
-	}
-
-	v.HomeDir = usr.HomeDir
+	home, err := homedir.Dir()
+	utils.ErrorExit(err)
+	v.HomeDir = home
 	v.TmuxDir = path.Join(v.HomeDir, ".tmux.d")
 	v.ProjTypeConfigDir = path.Join(v.HomeDir, ".tmux-project")
