--- constants.go
@@ -1,6 +1,0 @@
-package tmuxproject
-
-const (
-	ApplicationName string = "tmux-project"
-	Editor          string = "vim"
-)
