+++ utils/constants.go
@@ -1,0 +1,5 @@
+package utils
+
+const (
+	Editor          string = "vim"
+)
