+++ utils/strings.go
@@ -1,0 +1,11 @@
+package utils
+
+func StringInSlice(a string, list []string) bool {
+	for _, b := range list {
+		if b == a {
+			return true
+		}
+	}
+	return false
+}
+
